backend1/chatbot_service/requirements.txt
Flask
gunicorn
transformers
torch
scikit-learn
python-dotenv